+++ CHANGELOG.md
@@ -16,6 +16,7 @@ stack, where we shoe-horn in Vuetify as our main UI.
   customize their cron entry.
 - Sinks can now be searched/filtered based on ID and import/fetch
   status.
+- Added mass-operation tool for chunk selection
 
 ### Removed
 - Example dashboard (authenticated) and welcome (anonymous) scaffolding pages.